Feature/displaystorage

+++ app/src/main/java/uk/co/syski/client/android/data/entity/GPUEntity.java
@@ -25,5 +25,5 @@ public class GPUEntity {
 
     public String MemoryTypeName;
 
-    public int MemoryBytes;
+    public long MemoryBytes;
 }
